refactor(structure): Renamed store to providers
