+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from "next";
 import { League_Spartan } from "next/font/google";
 import { cookies } from "next/headers";
 import dynamic from "next/dynamic";
-import StoreProvider from "../store/StoreProvider";
-import AuthProvider from "../store/AuthProvider";
+import StoreProvider from "../providers/StoreProvider";
+import AuthProvider from "../providers/AuthProvider";
 import Navigation from "../components/Navbar/Navigation";
 import "./globals.css";
 import Footer from "../components/Footer/Footer";
